ElysiaStarter/requirements.txt
numpy>=2.1,<3           # Python 3.13 지원 버전
pygame>=2.6.1           # Python 3.13 호환 휠
pyquaternion==0.9.9     # pure-python
pyyaml>=6.0.2           # 3.13 호환
